+++ src/main/java/core/basesyntax/Main.java
@@ -1,0 +1,14 @@
+package core.basesyntax;
+
+public class Main {
+    private static final int NUMBER_OF_BALLS = 3;
+
+    public static void main(String[] args) {
+        Lottery lottery = new Lottery();
+        for (int i = 0; i < NUMBER_OF_BALLS; i++) {
+            Ball ball = lottery.getRandomBall();
+            System.out.println(ball);
+        }
+    }
+}
+
