--- src/main/java/core/basesyntax/main.java
@@ -1,11 +1,0 @@
- class Main {
-    private static final int NUMBER_OF_BALLS = 3;
-
-    public static void main(String[] args) {
-        Lottery lottery = new Lottery();
-        for (int i = 0; i < NUMBER_OF_BALLS; i++) {
-            Ball ball = lottery.getRandomBall();
-            System.out.println(ball);
-        }
-    }
-}
